[all] rename SimpleAsymmetricZOI to AsymmetricZOI
