+++ Benchmarks/ExampleSetups/ExmouthGulf/setup_pymanga.xml
@@ -2,7 +2,7 @@
 <MangaProject>
     <resources>
         <aboveground>
-            <type> SimpleAsymmetricZOI </type>
+            <type> AsymmetricZOI </type>
             <domain>
                 <x_1> 0 </x_1>
                 <y_1> 0 </y_1>
