+++ Benchmarks/ExampleSetups/FirstApplication/example_setup.xml
@@ -2,7 +2,7 @@
     <random_seed>643879</random_seed>
     <resources>
         <aboveground>
-            <type> SimpleAsymmetricZOI </type>
+            <type> AsymmetricZOI </type>
             <domain>
                 <x_1> 0 </x_1>
                 <y_1> 0 </y_1>
