+++ Benchmarks/ModuleBenchmarks/PlantModules/Bettina/Aboveground/AsymmetricZOI/AsymmetricZOI.xml
@@ -2,7 +2,7 @@
     <random_seed>643879</random_seed>
     <resources>
         <aboveground>
-            <type> SimpleAsymmetricZOI </type>
+            <type> AsymmetricZOI </type>
             <domain>
                 <x_1> 0 </x_1>
                 <y_1> 0 </y_1>
@@ -17,7 +17,7 @@
         </belowground>
     </resources>
     <plant_dynamics>
-        <type> Kiwi </type>
+        <type> Bettina </type>
         <mortality>NoGrowth</mortality>
     </plant_dynamics>
     <population>
@@ -27,7 +27,7 @@
             <distribution>
                 <type> GroupFromFile </type>
                 <n_recruitment_per_step> 0 </n_recruitment_per_step>
-                <filename> Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/initial_population.csv </filename>
+                <filename> Benchmarks/ModuleBenchmarks/PlantModules/Bettina/ag_initial_population.csv </filename>
             </distribution>
         </group>
     </population>
@@ -47,7 +47,9 @@
         <allow_previous_output>True</allow_previous_output>
         <output_dir>Benchmarks/TestOutputs/</output_dir>
         <geometry_output> r_stem </geometry_output>
-        <geometry_output> height </geometry_output>
+        <geometry_output> h_stem </geometry_output>
+        <geometry_output> r_crown </geometry_output>
+        <geometry_output> r_root </geometry_output>
         <growth_output> growth </growth_output>
         <growth_output> ag_resources </growth_output>
         <growth_output> bg_resources </growth_output>
