+++ Benchmarks/ModuleBenchmarks/PlantModules/Bettina/Aboveground/AsymmetricZOI/ReferenceFiles/AsymmetricZOI_CI/Population_t_0002000000.0.csv
